Renaming package to `unified_analytics` (#19)
* Global search and replace for dash_analytics -> unified_analytics

* Changing filenames and directories

* Clean up to pubspec + README

* Updating .github files

* Removing dash from md files + adding more detail to enums

* Fixing error in usage guide
