--- .github/ISSUE_TEMPLATE/dash_analytics.md
@@ -1,5 +1,0 @@
----
-name: "package:dash_analytics"
-about: "Create a bug or file a feature request against package:dash_analytics."
-labels: "package:dash_analytics"
----
