+++ .github/ISSUE_TEMPLATE/unified_analytics.md
@@ -1,0 +1,5 @@
+---
+name: "package:unified_analytics"
+about: "Create a bug or file a feature request against package:unified_analytics."
+labels: "package:unified_analytics"
+---
