+++ CODEOWNERS
@@ -2,4 +2,4 @@
 # for code within the repository
 # https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners
 
-/pkgs/dash_analytics/ @eliasyishak
+/pkgs/unified_analytics/ @eliasyishak
