--- pkgs/dash_analytics/README.md
@@ -1,15 +1,0 @@
-[![package:dash_analytics](https://github.com/dart-lang/tools/actions/workflows/dash_analytics.yml/badge.svg)](https://github.com/dart-lang/tools/actions/workflows/dash_analytics.yml)
-
-## What's this?
-
-This package is intended to be used on Dash (Flutter, Dart, etc.) related
-tooling only. It provides APIs to send events to Google Analytics using the
-Measurement Protocol.
-
-This is not intended to be general purpose or consumed by the community. It is
-responsible for toggling analytics collection for Dash related tooling on each
-developer's machine.
-
-## Using this package
-
-Refer to the [usage guide](USAGE_GUIDE.md).
