+++ pkgs/unified_analytics/lib/src/constants.dart
