+++ pkgs/unified_analytics/lib/src/initializer.dart
