+++ pkgs/unified_analytics/lib/src/log_handler.dart
