+++ pkgs/unified_analytics/lib/src/session.dart
