+++ pkgs/unified_analytics/pubspec.yaml
@@ -1,11 +1,11 @@
-name: dash_analytics
+name: unified_analytics
 description: >-
-  A package for logging analytics for all dash related tooling to Google
-  Analytics.
+  A package for logging analytics for all Dart and Flutter related tooling
+  to Google Analytics.
 # When updating this, keep the version consistent with the changelog and the
 # value in lib/src/constants.dart.
 version: 0.1.0-dev
-repository: https://github.com/dart-lang/tools/tree/main/pkgs/dash_analytics
+repository: https://github.com/dart-lang/tools/tree/main/pkgs/unified_analytics
 
 environment:
   sdk: '>=2.19.0 <3.0.0'
